requirements.txt
absl-py==2.1.0
addict==2.4.0
alabaster==1.0.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
autoflake==2.3.1
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm @ file:///croot/comm_1709322850197/work
commonmark==0.9.1
contourpy @ file:///work/perseverance-python-buildout/croot/contourpy_1701756524386/work
crcmod==1.7
cryptography==43.0.1
cupy-cuda12x==13.3.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.33
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
-e git+https://github.com/facebookresearch/detectron2.git@70f454304e1a38378200459dd2dbca0f0f4a5ab4#egg=detectron2&subdirectory=../../detectron2
-e git+https://github.com/IDEA-Research/detrex.git@c56d32e3d0262cff9835ebe80a0642965ae0cb3e#egg=detrex
docker-pycreds==0.4.0
docutils==0.21.2
easydict==1.13
einops==0.8.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fairscale==0.4.13
fastrlock==0.8.2
filelock==3.14.0
flake8==7.1.1
flatbuffers==24.3.25
fonttools @ file:///croot/fonttools_1713551344105/work
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.25.1
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.9
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1718287989724/work
isort==5.13.2
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jmespath==0.10.0
jupyter_client @ file:///work/perseverance-python-buildout/croot/jupyter_client_1701731788627/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
keras==3.5.0
kiwisolver @ file:///work/perseverance-python-buildout/croot/kiwisolver_1698847502605/work
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib==3.9.2
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mccabe==0.7.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
ml-dtypes==0.4.1
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.5
model-index==0.1.11
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1720002482208/work
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.12.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1721059439630/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy==1.5.0
ply @ file:///work/perseverance-python-buildout/croot/ply_1698846188635/work
portalocker==2.10.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.4
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pybind11==2.13.6
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pyflakes==3.2.0
Pygments==2.18.0
pyparsing @ file:///croot/pyparsing_1725041624202/work
PyQt5==5.15.10
PyQt5-sip @ file:///work/perseverance-python-buildout/croot/pyqt-split_1698847927472/work/pyqt_sip
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.3.3
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz==2023.4
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
recommonmark==0.7.1
requests==2.28.2
rich==13.4.2
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
setuptools==75.2.0
shapely==2.0.6
sip @ file:///work/perseverance-python-buildout/croot/sip_1698846977033/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
submitit==1.5.2
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2024.8.30
timm==1.0.9
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.65.2
traitlets @ file:///croot/traitlets_1718227057033/work
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3==1.26.20
wandb==0.18.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.0.4
wheel==0.43.0
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.1
zipp==3.20.2
